Convert `query_graph::extract_subgraphs_from_supergraph` into `supergraph` with split-up submodules
